startup: Only remain running at startup if a receiver is found
This changes the .desktop "startup" file so that if this is installed,
solaar will shutdown after initial startup when no receivers are
present. This saves a lot of memory, and removes the icon from view.
This is quite useful for those who dock their mouse, and/or only use it
occasionally. This should to be changed in solaar, because otherwise it
requires overwriting a system file which should ideally only be removed
via package uninstallation, and having this installed for when it is
needed is very useful.