docs: Fix punctuation and language in rules.md (#2032)
* Fix punctuation and language in rules.md

Fixed a few cases of backquote mistakes, unescaped <>, and misc language typos.

Took the liberty to unitalicise the third paragraph. Obviously it's important, but an entire paragraph in italics is hard to read.

I noticed there were two slightly different paragraphs about Setting conditions, which I assume is a mistake. I don't know which is better than the other, so I simply moved them next to each other for ease of comparison and manual merging/editing.

* Fixed erroneous backquote in rules.md

Fixed a ` which should have been a <. Error introduced by myself in the first place!